テスト間でFakeNotificationUsageDaoを共通化してコード重複を削減
