+++ frontend/android/feature-notification-usage/src/test/kotlin/net/matsudamper/money/frontend/android/feature/notificationusage/NotificationUsageRepositoryAndroidImplTest.kt
@@ -1,9 +1,6 @@
 package net.matsudamper.money.frontend.android.feature.notificationusage
 
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.first
-import kotlinx.coroutines.flow.map
 import io.kotest.core.spec.style.DescribeSpec
 import io.kotest.matchers.shouldBe
 import io.kotest.matchers.shouldNotBe
@@ -18,7 +15,7 @@ public class NotificationUsageRepositoryAndroidImplTest : DescribeSpec(
     {
         describe("通知利用履歴 repository") {
             it("未追加は parser に一致した通知だけ出る") {
-                val dao = RepositoryFakeNotificationUsageDao(
+                val dao = FakeNotificationUsageDao(
                     initialEntities = listOf(
                         NotificationUsageEntity(
                             notificationKey = "matched",
@@ -60,7 +57,7 @@ public class NotificationUsageRepositoryAndroidImplTest : DescribeSpec(
             }
 
             it("通知は受信時刻の降順で出る") {
-                val dao = RepositoryFakeNotificationUsageDao(
+                val dao = FakeNotificationUsageDao(
                     initialEntities = listOf(
                         NotificationUsageEntity(
                             notificationKey = "old-not-added",
@@ -105,7 +102,7 @@ public class NotificationUsageRepositoryAndroidImplTest : DescribeSpec(
             }
 
             it("追加済みに更新できる") {
-                val dao = RepositoryFakeNotificationUsageDao(
+                val dao = FakeNotificationUsageDao(
                     initialEntities = listOf(
                         NotificationUsageEntity(
                             notificationKey = "matched",
@@ -132,7 +129,7 @@ public class NotificationUsageRepositoryAndroidImplTest : DescribeSpec(
             }
 
             it("値が変わった同じ Android 通知は別通知として保存する") {
-                val dao = RepositoryFakeNotificationUsageDao(
+                val dao = FakeNotificationUsageDao(
                     initialEntities = listOf(
                         NotificationUsageEntity(
                             notificationKey = "matched",
@@ -173,7 +170,7 @@ public class NotificationUsageRepositoryAndroidImplTest : DescribeSpec(
             }
 
             it("値が同じ Android 通知の再取り込みでは保存済み通知を上書きしない") {
-                val dao = RepositoryFakeNotificationUsageDao(
+                val dao = FakeNotificationUsageDao(
                     initialEntities = listOf(
                         NotificationUsageEntity(
                             notificationKey = "matched",
@@ -211,68 +208,6 @@ public class NotificationUsageRepositoryAndroidImplTest : DescribeSpec(
         }
     },
 )
-
-private class RepositoryFakeNotificationUsageDao(
-    initialEntities: List<NotificationUsageEntity>,
-) : NotificationUsageDao {
-    private val entitiesFlow = MutableStateFlow(initialEntities)
-
-    override fun observeAll(): Flow<List<NotificationUsageEntity>> {
-        return entitiesFlow.map { entities ->
-            entities.sortByReceivedAtDescending()
-        }
-    }
-
-    override fun observeNotAdded(): Flow<List<NotificationUsageEntity>> {
-        return entitiesFlow.map { entities ->
-            entities.filter { it.isAdded.not() }
-                .sortByReceivedAtDescending()
-        }
-    }
-
-    override fun observeAdded(): Flow<List<NotificationUsageEntity>> {
-        return entitiesFlow.map { entities ->
-            entities.filter { it.isAdded }
-                .sortByReceivedAtDescending()
-        }
-    }
-
-    override suspend fun insert(entity: NotificationUsageEntity) {
-        if (entitiesFlow.value.none { it.notificationKey == entity.notificationKey }) {
-            entitiesFlow.value += entity
-        }
-    }
-
-    override suspend fun findByKey(notificationKey: String): NotificationUsageEntity? {
-        return entitiesFlow.value.firstOrNull { it.notificationKey == notificationKey }
-    }
-
-    override fun observeByKey(notificationKey: String): Flow<NotificationUsageEntity?> {
-        return entitiesFlow.map { entities ->
-            entities.firstOrNull { it.notificationKey == notificationKey }
-        }
-    }
-
-    override suspend fun markAsAdded(notificationKey: String, moneyUsageId: Int?) {
-        entitiesFlow.value = entitiesFlow.value.map { entity ->
-            if (entity.notificationKey == notificationKey) {
-                entity.copy(
-                    isAdded = true,
-                    moneyUsageId = moneyUsageId,
-                )
-            } else {
-                entity
-            }
-        }
-    }
-
-    private fun List<NotificationUsageEntity>.sortByReceivedAtDescending(): List<NotificationUsageEntity> {
-        return sortedWith(
-            compareByDescending<NotificationUsageEntity> { it.receivedAtEpochMillis }
-                .thenByDescending { it.postedAtEpochMillis },
-        )
-    }
-}
 
 private class RepositoryComExampleParser : NotificationUsageParser {
     override val filterDefinition: NotificationUsageFilterDefinition = NotificationUsageFilterDefinition(
